+++ packages/terra-clinical-item-collection/src/_TableHeaderCell.jsx
@@ -20,7 +20,7 @@ const TableHeaderCell = ({ columnType, ...customProps }) => {
   if (columnType === 'accessory') {
     return (<th {...customProps} className={cx(`column-${columnType}`)} />);
   }
-  return (<th {...customProps} data-class={cx(`column-${columnType}`)} />);
+  return (<th {...customProps} data-terra-clinical-item-collection-column={`${columnType}`} />);
 };
 
 TableHeaderCell.propTypes = propTypes;